Attempt to move around hook wrapping to hooks file
ExampleGroup delegates the running of before and after hooks to the Hooks
module. This is an aborted attempt to do the same for around hooks.

My intention was to have an AroundHook take a wrapped example and further
itself around it and return the result. Whilst this had an element of
increased consistency, it didn't make the code any easier to follow.

As the before/after hook refactoring may be desirable, I'm committing those
changes only.