Fix a bug with before/after(:all) hooks from unfiltered/excluded example groups.
Previously, before/after(:all) hooks in an example group with a nested example group would always be run, even if none of the specs in the example group were run due to inclusion/exclusion filtering.  We should only run the before/after(:all) hooks if there are any corresponding examples to run.

- Closes #161.