Fix confusing error message in `be_within`
This is a patch for the following issue:

https://github.com/rspec/rspec-expectations/issues/92

It is possible to have duck typing for numerical operations, but the minus
sign is often used for other operations as well such as the set
difference that Array#- performs. In that case, even if it does not fail
on the :- method, it will probably fail on the subsequent :abs and :<=
calls.