Remove the automatic call to RSpec.reset.
As far as I (Myron) can tell, the reset doesn't
provide any useful value to folks running specs
the normal way via the `rspec` command, and it
takes a bit of extra time to do so. It also limits
flexibility as alternate runners can't prevent RSpec
from calling this if they do not want it, but by
us no longer calling it, they can now decide if they
want things reset and can call it at the appropriate time.