Fix a `described_class` edge case.
In a case like:

  describe "in some context" do
    # example { described_class }
    describe MyClass do
      example { described_class }
    end
  end

...the value of `described_class` in the inner example group
changed based on whether or not the example in the outer group
was commented out or not. This was due to the fact that
`GroupMetadataHash#described_class` would return the parent
group's `described_class` if the key was present in that
group's metadata hash, but given the fact that `described_class`
is lazily computed and stored, the key would be present in the
parent group's metadata hash if and only if `described_class`
was called on the parent group--which would only be done if
the outer example was defined.

By adding the extra guard against nil, it allows the normal logic
to run to find the example group:

- scan the groups from the outer to the inner
- return the first non-string/non-symbol `describe` arg