Prevent `the examples should all pass` cuke step false positives.
It did not actually verify that any examples were run, and lead
to a false positive cucumber scenario discovered by @shepmaster
in #769. The underlying assumption of "the examples should all pass"
is that there were actually examples.

I think "the process should succeed even though no examples were run"
is a better description of the case when there are no examples.