Default `output_stream` to $stdout.
This is what the docs say we do. It's also necessary
so that if `configuration.reporter` is accessed before
the CommandLine class has set the output_stream, it
uses $stdout as its output rather than a `StringIO.new`.

Before, issuing a deprecation from `option_parser.rb`
was causing no spec output at all since that happened
before the output_stream was set.