Rename `parse_options` to `options` and have it return a memoized value.
The old way was odd. It created an ordering dependency
(you had to call `parse_options` after instantiating the
object but before doing anything else) and calling it a second
time had the side effect of re-parsing the args, which wasn't
needed or intended.

It's far easier/better to make `options` do the parsing
and memoize the result.