Refine `described_class` inference further.
In the last commit, I made inner `describe SomeClass`
example groups get `SomeClass` as the described_class.
The logic confusingly allowed a `:described_class` metadata
key in an outer group take precedence, which meant that the
behavior depended on whether or not `described_class` was
referenced from the outer group.

This fixes that, making it work consistently:

- Use `:described_class` is present.
- Otherwise, use the group's `describe` arg
  (unless it's something like a string).
- If neither of those, look in the parent and
  recurse.