Remove code that wasn't doing anything.
The `AT_EXIT_HOOK_BACKTRACE_LINE` was wrong now,
anyway -- the hook was more than 2 lines before
the constant declaration. So it was no longer
serving it's purpose. I can't repro the original
bug now, anyway, and it was related to our `at_exit`
hook which is no longer used when running specs
via the `rspec` command.