Specify relationship of `after` hooks and mock verification.
While it’s not recommended to rely on this behavior
(it's pretty weird to satisfy a mock expectation
in an `after` hook...), this is consistent with how
we treat `after` hooks: we don't set the example's
`execution_status` until after the `after` hooks run
because if an error happens in them we need to set
the examples's status to `:failed`. Thus, we should
not verify mock expectations until after `after` hooks
since the example has not fully completed until
those hooks have run.

This clarifies the behavior in question in #1769.