Fix intermittently failing specs.
These specs would fail when run in isolation but
would pass when run after other specs that used the
`match` and `include` matchers first (causing them
to be loaded). To make them consistently pass,
we need to get the matchers before manipulating
the load path to ensure they are already loaded.