Fix `RSpec::Core::RakeTask#failure_message`.
`system` returns a boolean value to indicate success/failure,
but the use of `failure_message` was in a `rescue` block that
never got executed.  It appears this has been broken since
ea70e4edbbc52846dcd63308ee349166bbcef29c. Before that commit,
we used `Rake::FileUtilsExt#ruby`, which does indicate failure
by raising an error (and thus the `rescue` was correct). In
ea70e4edbbc52846dcd63308ee349166bbcef29c, we switched to using
`system` and the rescue was left in place but never got hit
anymore.

The lack of test coverage here is why we never noticed, so I
addressed that as well.