run suite hooks in an example group context
All other types of hooks are run in the context of example group
during an individual example group or examples run (respectively),
as a suite hook uses a special special sub class of example to run in
it wasn't even being setup correctly with an example group so they ran
in the context of nil. fixed.