Include current error message in raise_error warning
Fix for issue #822

https://github.com/rspec/rspec-expectations/issues/822

When the raise_error() matcher triggers a warning it does not include
the cause of the actual error. This commit changes the warning message
to include the actual error that was raised.

I've moved setting the warning message `RaiseError#warn_about_bare_error`
further down in `RaiseError##matches` because the warning message needs
an instance variable to be set in order to display the current error.