Deal with circular dependency between world and config.
`RSpec::World#initialize` references `RSpec.configuration`
and `RSpec::Configuration` references `RSpec.world` in a few
places. This circular dependency is not ideal and has triggered
infinite recursion during some recent changes I've tried to make.
It would be nice to get rid of it entirely, but for now, this is
the simplest change I can make it solve my immediate problems.