Ensure groups and examples are registered before invoking config definition hooks.
This is important so that if you load any support files that
define hooks from within a definition hook, it can apply the
hooks to the already-defined example groups (as is now
supported via #2189). Without this change, the current group
being defined is not eligible to have newly loaded hooks
applied to it.