Get failure message from `exception_presenter`
Currently a few of the formatters get their failure message directly from the
exception[1].

This commit changes the HTML formatter and gets the failure message from the
`FailedExampleNotification#message_lines`, which delegates this message to the
`ExceptionPresenter#message_lines`. Meaning we the HTML formatter uses the
exception presenter to get the exception.

1.) https://github.com/rspec/rspec-core/blob/master/lib/rspec/core/formatters/html_formatter.rb#L71