Handle errors during spec file load time in a nicer manner.
- Format them like other errors.
- Allow other files to be loaded so you can see all
  the load-time errors.
- Abort the test suite before running any examples.

This has a slight affect on the bisect server: previously
it would get no run results at all if a load-time error
occurred in the spec files. Now it gets run results but
they are empty. The change happens because the error is handled,
allowing the bisect formatter to report results in `start_dump`.