Add support for $XDG_CONFIG_HOME/rspec/options
This is a XDG Base Directory Specification compatible alternative to
~/.rspec.

~/.rspec has higher precedence in order to reduce risk in cases where
users have a file like this dormant. It is unlikely, but this way the
risk is lower.

If $XDG_CONFIG_HOME is not set, it will fall back to ~/.config, per the
specification.

Other changes:

The "isolated home" example tag has been extended with a
`create_fixture_file` helper that also takes care to create the
directory where the file resides in, if it does not already exist.
Other file creations have been migrated to this method for consistency.

The method is only available when an example is tagged with
`:isolated_home => true`, for safety.

In case the developer running these tests have a custom $XDG_CONFIG_HOME
set, it is cleared out when using an isolated home so their real files
are not touched.